agent-utilities/requirements.txt
pytest
pylint
openmeteo-requests
requests-cache
retry-requests
numpy
pandas
geopy
dateparser
pydantic
fastmcp
fastapi
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
